requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
Django==1.9.5
django-phonenumber-field==1.1.0
djangorestframework==3.3.3
httpie==0.9.3
pexpect==4.0.1
pickleshare==0.7.2
ptyprocess==0.5.1
Pygments==2.1.3
psycopg2
requests==2.9.1
simplegeneric==0.8.1
traitlets==4.2.1
